requirements.txt
streamlit>=1.37.0
requests>=2.31.0
pandas>=2.2.2
plotly>=5.23.0
openpyxl>=3.1.5
XlsxWriter>=3.2.0
psycopg2-binary>=2.9.9
bcrypt>=4.1.3
python-dotenv>=1.0.1
